fix: display explicit error message when model not downloaded
For a couple of functionalities, we need to load a spacy model that is not included by default. The developer has to download it manually in addition to the package. If the model is not present locally, an error will occur. The message is not meaningful. This commit will make it more explicit.